Break out MainPanel.tsx into App and GameBoard
Deleted MainPanel.tsx, this was a holdover from the auto-gen'd create-react-app and didn't make sense for 2048.
Broke out the logic it had into App.tsx which now has all the core title and scaffolding,
and GameBoard.tsx which is now singly responsible for knowing how to render the game board itself.